backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
numpy
pandas
scikit-fuzzy